Update GUI_Test.py
I finished adding the remainder of the column for data cleaning, I also made the title and text more formal, I also began to add a function for finding the  Mean. The code mostly worked as intended, the title was changed and data cleaned correctly. Now the GUI closes upon ending the cleaning opening a new GUI for analysis, the Mean function is currently under development and was not expected to function entirely as intended, right now it creates a whole new column for one mean which is not desirable.
The plan is to finish making the Mean function and then copying it and altering it slightly to fit all other common statistical functions, after that  I plan to create yet another GUI for graph options before finally making the GUIs more aesthetically pleasing.