Renamed to Main.py, Update Main.py
Add drop down boxes in the data analysis process for ease of adding the statistical values, created functions for a vast majority of statistical values for future use, made code that checked if certain numerical columns were in the data as to not break the code. Several tests were required to fix small errors such as square brackets instead of round ones but eventually the drop down boxes worked as intended. Next is to add the remaining statistical functions and make them useable.